server_code/requirements.txt
google-generativeai==0.8.5
google-genai
